Task 3

+++ 0x10-variadic_functions/3-print_all.c
@@ -25,10 +25,10 @@ void print_all(const char * const format, ...)
 					printf("%s%c", sep, va_arg(list, int));
 					break;
 				case 'i':
-					printf("%s%f", sep, va_arg(list, int));
+					printf("%s%d", sep, va_arg(list, int));
 					break;
 				case 'f':
-					printf("%s%d", sep, va_arg(list, int));
+					printf("%s%f", sep, (double) va_arg(list, int));
 					break;
 				case 's':
 					str = va_arg(list, char *);
